+++ src/main/java/kuit/project/beering/domain/image/DrinkImage.java
@@ -1,10 +1,12 @@
-package kuit.project.beering.domain;
+package kuit.project.beering.domain.image;
 
 import jakarta.persistence.*;
+import kuit.project.beering.domain.Drink;
+import kuit.project.beering.domain.image.Image;
 
 @Entity
 @DiscriminatorValue("drink")
-public class DrinkImage extends Image{
+public class DrinkImage extends Image {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "drink_id")
